refactor: 카드 초기 설정 및 등록 로직 커스텀 훅으로 분리
